# RGB to/from Gray rewritten to wide intrinsics