… modules/imgproc/src/opencl/color_rgb.cl