@savuor 2018-12-13
@alalek ARM builds fixed